remove flow plugins and rules configuration
We've been using TypeScript on ocw-studio for a bit now, and in a few
other places as well. It's been nice so far, and it's not looking like
we're likely to use Flow on projects going forward, although we are
still using it on micromasters and open-discussions.

Right now the current config requires installing eslint-plugin-flow-vars
and eslint-plugin-flowtype even if the project is using TypeScript,
which is a bit annoying. So I thought, well, we're not going to use Flow
on new projects going forward, so we should go ahead and remove these
rules.

pr #12